Refactor player and physics modules
