+++ environment.js
@@ -1,7 +1,8 @@
 // Import Three.js
 import * as THREE from 'three';
 
-import { collidableObjects, addToScene } from './main.js';
+import { collidableObjects } from './physics.js';
+import { addToScene } from './main.js';
 
 class Wall {
     constructor({x=0, y=2.5, z=-5, w=6, h=5, d=1} = {}){
